Extract some assertion groups into traits
